Changes made to project
